Change to use single test results file for jsc tests https://bugs.webkit.org/show_bug.cgi?id=264908 rdar://118482056
Reviewed by Ryan Haddad.

Currently we generate a test status file for each jsc test we run, and then
we will use a find command for remote devices to get the content of the each
file.

This find command is time consuming, it may take forever for low end embedded
device. This patch will make the test result ouput to a single file, then we copy
the file back the host, open it locally, and process the results line by line.
It will save a lot of time for the script to report the final status.

* Tools/Scripts/run-jsc-stress-tests:
* Tools/Scripts/webkitruby/jsc-stress-test-writer-default.rb:
* Tools/Scripts/webkitruby/jsc-stress-test/executor.rb:

Canonical link: https://commits.webkit.org/272663@main