Optimize SVGRenderSupport::layoutDifferentRootIfNeeded() https://bugs.webkit.org/show_bug.cgi?id=267014 rdar://120392477
Reviewed by Said Abou-Hallawa.

SVGRenderSupport::layoutDifferentRootIfNeeded() shows up on profiles of the Suits
MotionMark subtest.

We only need to call findTreeRootObject() on the renderer if the there are any  clipper, masker,
filters etc are present, so pass the RenderElement which is the client of the SVG resources,
and lazily call findTreeRootObject() on it when necessary.

Also check m_clipperFilterMaskerData and m_markerData before calling the lambda on their
contents.

* Source/WebCore/rendering/svg/SVGRenderSupport.cpp:
(WebCore::SVGRenderSupport::layoutDifferentRootIfNeeded):
* Source/WebCore/rendering/svg/SVGResources.cpp:
(WebCore::SVGResources::layoutDifferentRootIfNeeded):
* Source/WebCore/rendering/svg/SVGResources.h:

Canonical link: https://commits.webkit.org/272696@main