[WebGPU] Texture sample type is unnecessarily generated against auto-generated layouts https://bugs.webkit.org/show_bug.cgi?id=267093 <radar://120487902>
Reviewed by Tadeu Zagallo.

The sample type of textures for auto-generated layouts is always a constant,
so it is not possible to validate them during bind group creation.

This was observed running the resource_init validation suite, which is currently
not running on EWS due to timeouts.

* Source/WebGPU/WebGPU/BindGroup.mm:
(WebGPU::Device::createBindGroup):

Canonical link: https://commits.webkit.org/272709@main