Revert removal of else branch in SelectorChecker::matchRecursively() https://bugs.webkit.org/show_bug.cgi?id=267184
Reviewed by Tim Nguyen.

272726@main (re-landed as 272729@main) removed the else branch under
the assumption that all paths had early returns. But that is not the
case. E.g., a user agent shadow root containing an element with a
matching pseudo attribute.

While I'm not sure how to cover this with a test and no tests
regressed, it seems best to restore the status quo prior to that
commit.

* Source/WebCore/css/SelectorChecker.cpp:
(WebCore::SelectorChecker::matchRecursively const):

Canonical link: https://commits.webkit.org/272738@main