[css-flexbox] Start storing flex line structure as the output of flex layout. https://bugs.webkit.org/show_bug.cgi?id=267006 rdar://problem/120390262
Reviewed by NOBODY (OOPS!).

This is a precursor to various bugs related to determining the baselines
of flexboxes. The spec determines the flex item that will be used to
best represent the baseline of flexbox using the structure of the flex
lines. At that point we no longer have the knowledge of the structure of
the flex lines so we should start keeping track of them at the end of
layout.

Unfortunately reconstructing the flex lines is not a trivial process
because:

1.) Creating a flex item via constructFlexItem is not a const operation
so we risk mutating the renderers
2.) Attempting to create the flex lines by querying the renderers
directly and checking the flexbox's style is also error prone. Any
oversights could result in an invalid structure of the flex line

* Source/WebCore/rendering/RenderFlexibleBox.cpp:
(WebCore::RenderFlexibleBox::layoutBlock):
(WebCore::RenderFlexibleBox::layoutFlexItems):
* Source/WebCore/rendering/RenderFlexibleBox.h: