[WPE] WPE Platform: wrong usage reported in DMABuf buffer formats under wayland https://bugs.webkit.org/show_bug.cgi?id=267406
Reviewed by Alejandro G. Castro.

When there's more than one tranche, the flags for the next tranches are
added to the previous one. This is because we are using std::move with
the Tranche struct without defining a move constructor, so the flags
member is not set to 0 after the move.

* Source/WebKit/WPEPlatform/wpe/wayland/WPEViewWayland.cpp:
(DMABufFeedback::Tranche::Tranche): Define a move constructor and make it non copyable.

Canonical link: https://commits.webkit.org/272911@main