[Perf dashboard] cost should be an array of Float32Array in findOptimalSegmentationInternal https://bugs.webkit.org/show_bug.cgi?id=267434
Reviewed by Dewei Zhu.

The cost array in findOptimalSegmentationInternal is always intended to be an array of Float32Array
but we were forgetting to instantiate Float32Array for the very first item in the array.

This PR makes it more consistent across the array.

* Websites/perf.webkit.org/public/shared/statistics.js:
(Statistics.new.findOptimalSegmentationInternal):

Canonical link: https://commits.webkit.org/272948@main