Scroll To Text Fragment Text Directives don't find text with additional unrendered white space in their node data. https://bugs.webkit.org/show_bug.cgi?id=267460 rdar://120913588
Reviewed by Ryosuke Niwa.

The spec for Scroll to Text Fragment says to just look a the .data() on the node
to determine it's text content and search for the fragment directive in the page
text. However, this does not account for extra white space that is removed when rendering.
Therefore, we should use TextItterator which does account for a remove this unrendered white
space. This is technically against the spec, but other engines similarly ignore this
extraneous white space when searching, so the spec should be updated accordingly.
If TextItterator does not find the text, we fall back to the original spec way of finding
the text.

* LayoutTests/http/tests/scroll-to-text-fragment/whitespace-collapse-expected.html: Added.
* LayoutTests/http/tests/scroll-to-text-fragment/whitespace-collapse.html: Added.
* Source/WebCore/dom/FragmentDirectiveRangeFinder.cpp:
(WebCore::FragmentDirectiveRangeFinder::findRangeFromNodeList):

Canonical link: https://commits.webkit.org/273016@main